requirements.txt
certifi==2019.6.16
chardet==3.0.4
cron-descriptor==1.2.21
croniter==0.3.30
idna==2.8
oauthlib==3.0.2
PySocks==1.7.0
python-crontab==2.3.8
python-dateutil==2.8.0
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
toml==0.10.0
tweepy==3.8.0
urllib3==1.25.3
